Add TrustedFS.
Add the TrustedFS type, representing an fs.FS in a safe way.

The only ways to create a TrustedFS is from an embed.FS, or from a
TrustedSource. This ensures that a TrustedFS is always under
application control.

This new safehtml API requires Go 1.16 or higher, since
that is when fs.FS was introduced. That seems reasonable, since the Go
team supports only the last two versions, and we are now on 1.17.

PiperOrigin-RevId: 405723011
Change-Id: I3f8554b25b74630386956ab240a4d0a70db0aee2